chore: Remove createPost controller and add post controller with create, remove, and update functionality
